tests

+++ tests/test_jobExecutor_6.sh
@@ -1,13 +1,21 @@
+
+if [ $# -ne 1 ]; then
+    echo "Usage: $0 <port>"
+    exit 1
+fi
+
+# Again without sleeps you wont see the expectet output
 killall progDelay
-jobCommander localhost 8000 issueJob progDelay 1000
-jobCommander localhost 8000 issueJob progDelay 1000
-jobCommander localhost 8000 issueJob progDelay 1000
-jobCommander localhost 8000 issueJob progDelay 1000
-jobCommander localhost 8000 issueJob progDelay 1000
-jobCommander localhost 8000 issueJob progDelay 1000
-jobCommander localhost 8000 setConcurrency 4
-jobCommander localhost 8000 poll
-jobCommander localhost 8000 stop job_4
-jobCommander localhost 8000 stop job_5
-jobCommander localhost 8000 poll
-jobCommander localhost 8000 exit
+jobCommander localhost $1 issueJob progDelay 10 &
+jobCommander localhost $1 issueJob progDelay 10 &
+jobCommander localhost $1 issueJob progDelay 10 &
+jobCommander localhost $1 issueJob progDelay 10 &
+jobCommander localhost $1 issueJob progDelay 10 &
+jobCommander localhost $1 issueJob progDelay 10 &
+# sleep 1
+jobCommander localhost $1 setConcurrency 4
+jobCommander localhost $1 poll
+jobCommander localhost $1 stop job_4
+jobCommander localhost $1 stop job_5
+jobCommander localhost $1 poll
+jobCommander localhost $1 exit
