+++ tests/test_jobExecutor_7.sh
@@ -1,4 +1,10 @@
+
+if [ $# -ne 1 ]; then
+    echo "Usage: $0 <port>"
+    exit 1
+fi
+
 killall progDelay
-jobCommander localhost 8000 issueJob progDelay 1000
-jobCommander localhost 8000 stop job_2
-jobCommander localhost 8000 exit
+jobCommander localhost $1 issueJob progDelay 10 &
+jobCommander localhost $1 stop job_2
+jobCommander localhost $1 exit
